+++ dropwizard-auth/src/main/java/io/dropwizard/auth/Authorizer.java
@@ -1,5 +1,7 @@
 package io.dropwizard.auth;
 
+import javax.annotation.Nullable;
+import javax.ws.rs.container.ContainerRequestContext;
 import java.security.Principal;
 
 /**
@@ -8,7 +10,6 @@ import java.security.Principal;
  * @param <P> the type of principals
  */
 public interface Authorizer<P extends Principal> {
-
     /**
      * Decides if access is granted for the given principal in the given role.
      *
@@ -16,5 +17,19 @@ public interface Authorizer<P extends Principal> {
      * @param role a user role
      * @return {@code true}, if the access is granted, {@code false otherwise}
      */
+    @Deprecated
     boolean authorize(P principal, String role);
+
+    /**
+     * Decides if access is granted for the given principal in the given role.
+     *
+     * @param principal a {@link Principal} object, representing a user
+     * @param role a user role
+     * @param requestContext a request context.
+     * @return {@code true}, if the access is granted, {@code false otherwise}
+     */
+    @SuppressWarnings("deprecation")
+    default boolean authorize(P principal, String role, @Nullable ContainerRequestContext requestContext) {
+        return authorize(principal, role);
+    }
 }