+++ dropwizard-configuration/src/test/java/io/dropwizard/configuration/EnvironmentVariableLookupTest.java
@@ -8,27 +8,17 @@ import static org.assertj.core.api.Assumptions.assumeThat;
 
 public class EnvironmentVariableLookupTest {
     @Test
-    public void defaultConstructorEnablesStrict() {
+    public void lookupThrowsExceptionInStrictMode() {
         assumeThat(System.getenv("nope")).isNull();
-
-        assertThatExceptionOfType(UndefinedEnvironmentVariableException.class).isThrownBy(()->
-            new EnvironmentVariableLookup().lookup("nope"));
+        assertThat(new EnvironmentVariableLookup().lookup("nope")).isNull();
     }
 
     @Test
     public void lookupReplacesWithEnvironmentVariables() {
-        EnvironmentVariableLookup lookup = new EnvironmentVariableLookup(false);
+        EnvironmentVariableLookup lookup = new EnvironmentVariableLookup();
 
         // Let's hope this doesn't break on Windows
         assertThat(lookup.lookup("TEST")).isEqualTo(System.getenv("TEST"));
         assertThat(lookup.lookup("nope")).isNull();
-    }
-
-    @Test
-    public void lookupThrowsExceptionInStrictMode() {
-        assumeThat(System.getenv("nope")).isNull();
-
-        assertThatExceptionOfType(UndefinedEnvironmentVariableException.class).isThrownBy(() ->
-            new EnvironmentVariableLookup(true).lookup("nope"));
     }
 }